fix: duplicate topic documentation
